+++ src/NodoItem.cpp
@@ -4,7 +4,7 @@ NodoItem::NodoItem(int  id){
    this->id =  id;
 }
 
-void  NodoItem::add_user(int rating,NodoUser* & new_user){
+void  NodoItem::add_user(float rating,NodoUser* & new_user){
     
     this->users[new_user] = rating;
 }
